More modification for the fakeable mode
Ignore-this: ee39fd77f5b994ae30e103d93debbafe

darcs-hash:20120203124759-5c877-32c704b81332e255f109d8b018bc340c11cdf5b1.gz

+++ Utils/sendall.sh
@@ -16,7 +16,7 @@ Send all the final states jobs to the cluster
 
 SYNTAX:
 
-   $0 [-r runperiod]  <WZ|WH> 
+   $0 [-r runperiod] [-F] <WZ|WH> 
 
 
    Note that the signal is a mandatory argument. 
@@ -29,6 +29,7 @@ SYNTAX:
 OPTIONS:
 
    [-r]: Set the Run period: 2011, 2011A, 2011B. Default: 2011
+   [-F]: Activate fakeable mode
 
 EOF
 }
@@ -54,11 +55,15 @@ if [ -z $1 ]; then
     exit;
 fi
 
-while getopts r:h o
+fakeable=""
+
+while getopts r:Fh o
 	do
 		case "$o" in
 			l)	runperiod=$OPTARG; 
 			 	shift;
+				shift;;
+			F)      fakeable="yes";
 				shift;;
 			h)	help;
 				exit 1;;
@@ -95,10 +100,30 @@ cd $PWD
 
 signal=$1
 
-echo "[sendall] Creating data files to extract root files"
+echo "[sendall] Info: Creating data files to extract root files"
 datamanagercreator -r $runperiod -f mmm;
-echo "[sendall] Warning: the ZJets_Madgraph sample is not yet implemented. Removing it"
-rm ZJets_Madgraph_datanames.dn;
+if [ -f ZJets_Madgraph_datanames.dn ]; then
+	echo "[sendall] Warning: the ZJets_Madgraph sample is not yet implemented. Removing it";
+	rm ZJets_Madgraph_datanames.dn;
+fi 
+if [ "$1" == "WZ" ]; then
+	echo "[sendall] Info: not needed the WH samples, removing";
+	rm WHTo*.dn;
+fi
+
+if [ "X"$fakeable == "X" ];
+then
+	echo "[sendall] Info: not needed the Fakes sample, removing";
+	rm Fakes*.dn;
+else
+	echo "[sendall] Info: not needed the Z+Jets, DY and TTbar samples, removing";
+	rm Z*_Powheg_datanames.dn;
+	rm DY*_Powheg_datanames.dn;
+	rm TTbar_Madgraph_datanames.dn;
+fi
+
+
+
 for finalstate in mmm mme eem eee;
 do 
 	i=$signal$finalstate
@@ -106,9 +131,14 @@ do
 	cd $i;
 	cp ../*.dn .
 	datamanagercreator Data -r $runperiod -f $finalstate;
-	datamanagercreator Fakes -r $runperiod -f $finalstate;
+	fakeoption=""
+	if [ "X"$fakeable == "Xyes" ];
+	then
+		datamanagercreator Fakes -r $runperiod -f $finalstate;
+		fakeoption="-F 3,2"
+	fi
 	echo "[sendall] Sending $finalstate -- Working directory: $i"; 
-	sendcluster submit -a $signal -f $finalstate -c MUON:../$cfgmmm,ELECTRON:../$cfgeee -F 3,2;
+	sendcluster submit -a $signal -f $finalstate -c MUON:../$cfgmmm,ELECTRON:../$cfgeee $fakeoption;
 	cd ../; 
 done
 rm *.dn
